fix(site): address review feedback (folder, status indicator, null guard, nit)
- Thread gitWatcher-preferred folder through to VSCodeMenuItem for
  correct monorepo folder resolution (C002)
- Remove early return null in WorkspacePill so the pill always renders
  when workspace data is available (C004)
- Remove stray {" "} whitespace in ChatTopBar (C005)
- Extract getWorkspaceStatusDisplay into shared
  WorkspaceStatusIndicator to deduplicate status icon/label logic
  between WorkspacePill and AgentChatPageView (C006)
